Merges SimpleTypingIndicator into TypingIndicator component.
Consolidates typing indicator components for maintenance.
Updates TypingIndicator component to handle animations.
